+++ sdk/resources/azure-resourcemanager-resources/src/main/java/com/azure/resourcemanager/resources/implementation/DeploymentStacksClientImpl.java
@@ -78,7 +78,7 @@ public final class DeploymentStacksClientImpl implements InnerSupportsGet<Deploy
      * proxy service to perform REST calls.
      */
     @Host("{$host}")
-    @ServiceInterface(name = "DeploymentStacksMana")
+    @ServiceInterface(name = "DeploymentStacksManagementClientDeploymentStacks")
     public interface DeploymentStacksService {
         @Headers({ "Content-Type: application/json" })
         @Get("/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Resources/deploymentStacks")